Hopefully the last macOS float conversion fix

+++ glfw/cocoa_time.c
@@ -42,7 +42,7 @@ void _glfwInitTimerNS(void)
     mach_timebase_info_data_t info;
     mach_timebase_info(&info);
 
-    _glfw.timer.ns.frequency = (info.denom * 1e9) / info.numer;
+    _glfw.timer.ns.frequency = (unsigned long long)((info.denom * 1e9) / info.numer);
 }
 
 
@@ -59,4 +59,3 @@ uint64_t _glfwPlatformGetTimerFrequency(void)
 {
     return _glfw.timer.ns.frequency;
 }
-
